+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## v1.34.3
 
+### Internal
 
+- [Netkan] Re-sort module properties after merging (#3957 by: HebaruSan)
 
 ## v1.34.2 (Minkowski²)